Add Create/List/Delete Namespaces APIs
This allows for namespaces to be created, listed and updated over the
gRPC API as well as via the IPAMer interface.

CreateNamespace is idempotent and can be called to ensure a namespace
exists before performing operations in it. The "root" namespace is
always created by default in all Storages.

ListNamespace checks the storage for namespaces and returns a list. The
same implementation is mostly used for checking namespaces internally
for other namespaced API operations.

DeleteNamespace is not idempotent, and will return an error if the
namespace to be deleted no longer exists.

Other implementation details included are:

- MonogoDB now uses a collections per namespace
- Error implementations were replaced with errors.New
- Tests were changed to use the require.NoError and require.ErrorsIs
  functions

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>